slim down the bloated mempool to prepare for an optimization
